9.5 final correction

+++ kodilla-good-patterns/src/main/java/food2Door/manufacturer/ProductInStock.java
@@ -1,13 +1,15 @@
 package food2Door.manufacturer;
 
+import java.math.BigDecimal;
+
 public class ProductInStock {
 
     private String productNameInStock;
     private String manufacturerName;
-    private double price;
+    private BigDecimal price;
     private int quantityInStock;
 
-    public ProductInStock(String productNameInStock, String manufacturerName, double price, int quantityInStock) {
+    public ProductInStock(String productNameInStock, String manufacturerName, BigDecimal price, int quantityInStock) {
         this.productNameInStock = productNameInStock;
         this.manufacturerName = manufacturerName;
         this.price = price;
@@ -22,14 +24,13 @@ public class ProductInStock {
         return manufacturerName;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
     public int getQuantityInStock() {
         return quantityInStock;
     }
-
 }
